chore: update landing page
